requirements.txt

Flask
gunicorn
pandas
numpy
xgboost
scikit-learn
joblib
python-dotenv
requests
psycopg2-binary
Flask-Cors